% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_bubble.R
\name{hierarchical_bubble}
\alias{hierarchical_bubble}
\title{A heirarchical bubble widget}
\usage{
hierarchical_bubble(data = list(), width = NULL, height = NULL,
  elementId = NULL)
}
\description{
A heirarchical bubble widget
}
