% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wide_to_long_hierarchy}
\alias{wide_to_long_hierarchy}
\title{Convert data into format required for d3 vis}
\usage{
wide_to_long_hierarchy(df, hierarchy_names, value_column)
}
\description{
Convert data into format required for d3 vis
}
